requirements.txt
discord.py>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
aiosqlite>=0.17.0
requests>=2.28.0
Pillow>=10.0.0
psutil>=5.9.0
